Extract preview URL core module
Move preview URL routing, state, and forwarding logic out of the
Sandbox Durable Object facade while keeping lifecycle ownership in the
core Sandbox class.
